Mapeamento completo

+++ Hospitalzinho/Entidades/Vacina.cs
@@ -6,7 +6,7 @@ namespace Hospitalzinho.Entidades
 {
     public class Vacina
     {
-        public virtual int Id { get; set; } // Identificador único do lote
+        public virtual long Id { get; set; } // Identificador único do lote
         public virtual string Lote { get; set; } // Lote da vacina
         public virtual DateTime DataProducao { get; set; }
         public virtual DateTime DataValidade { get; set; }
